server-Inherited_ACLs: Avoid using different sizes for mem_alloc and memcpy, some style cleanup.
It is a bit risky to use two different values for mem_alloc and memcpy - under some circumstances
these values probably don't match, leading to an invalid memory access. As the ACLs are accessed
without any special checks this shows that we'll also need additional protection in one of the
earlier patches.