Added patch to revert pixelformat changes (caused regression for many apps).
Well, usually we don't revert Wine commits in this branch. In this case the original author of the
patches suggested to revert them, because its unfinished and the rest of the patches will not be
accepted anymore in the near future. Without the complete patchset it causes a lot of side effects
(flickering, black/grey screen, missing video). This makes Wine basically useless for all affected
apps. Since the revert was not accepted upstream, we do it in our branch, until either the problems
are gone, or the revert got accepted.