winex11-Limited_Resolutions: Temporarily disable patchset.
The patchset itself should be fine, but statically linking against libXNVCtrl
might be problematic because of the different license. Moreover, on Debian it
is even a bit more complicated because a dynamic library is provided. Since we
do not want to have a fixed dependency on Nvidia-specific packages (winex11
would depend on it) we will have to investigate this a bit more...