Rebase against upstream Wine (what the ...).
Introduce a new 'advapi32-Revert_DACL' patchset for now, to revert all the
changes by Piotr Caban.  We have a completely independent and relatively
well-tested implementation of the same functionality in the wineserver.

TODO: Take a closer look at all the tests from our version and the proposed
change, and then finally decide which implementation we want to keep.