Removed patch to fix cross-compilation of 32-bit Wine on OSX.
On Clang 3.5 passing "-m32" implicitly enabled SSE2 support. In newer versions
of Clang this issue was fixed, so the patch has no effect anymore, and a
different solution is required to fix various build failures related to SSE2
support. Ideally Wine should automatically set it when compiling such
components, but for now we pass CFLAGS="-msse2" in the build scripts to work
around this bug.

In addition, there are several bugs in the cctools-port crosscompiling
toolchain - at several places host architecture and target architecture are
mixed up, leading to build failures when the toolchain was built for a 64-bit
host system and trying to compile for a 32-bit target. For our build servers,
we now apply patches to fix those issues properly. A quick&dirty workaround
would be to keep the CFLAGS="-m32".