updated xactengine-initial patchset
Include xact_class.h with <> to ensure the local file is found
over the one in the xaudio2_7 directory.

If one builds xaudio2_7 then one of the xactengine dll, a build
occurs because it's included the wrong header.

The build appears to work by chance because the xactengine dlls are
built before xaudio.