server-Desktop_Refcount: Remove patch set.
This was first added in 78db8b7950c18c7882f800449029c4b9662c25f0 with the
message "Added patch to fix possible leak of explorer.exe processes and
implement proper desktop refcounting." The message of course suggests that
even then no application was known that actually depended on this.

In practice this seems wrong, also: once all non-system processes other than
explorer have terminated, the only way explorer.exe can remain alive is if a
system process holds a reference to explorer. The only relevant system
processes are services, and services cannot display a GUI [this is enforced
since Vista], so this should never be the case.